Moves TextSource and TextSink to top level
